refactor(evm): restructure dirs for future developments
