+++ apps/evm/src/containers/VaultCard/VaultModal/OverviewTab/Overview/index.tsx
@@ -1,0 +1,11 @@
+import type { PendleVault } from 'types';
+
+import { PendleTotalDeposits } from './PendleTotalDeposits';
+
+export interface OverviewProps {
+  vault: PendleVault;
+}
+
+export const Overview: React.FC<OverviewProps> = ({ vault }) => {
+  return <PendleTotalDeposits vault={vault} />;
+};
