+++ apps/evm/src/containers/VaultCard/VaultModal/OverviewTab/StrategyDiagram/FlowArrow/index.tsx
