+++ apps/evm/src/containers/VaultCard/VaultModal/OverviewTab/StrategyDiagram/FlowNode/index.tsx
