+++ apps/evm/src/containers/VaultCard/VaultModal/OverviewTab/index.tsx
@@ -1,8 +1,8 @@
 import type { PendleVault } from 'types';
 
 import { MarketInfo } from './MarketInfo';
+import { Overview } from './Overview';
 import { StrategyDiagram } from './StrategyDiagram';
-import { TotalDeposits } from './TotalDeposits';
 
 export interface OverviewTabProps {
   vault: PendleVault;
@@ -11,7 +11,7 @@ export interface OverviewTabProps {
 export const OverviewTab: React.FC<OverviewTabProps> = ({ vault }) => {
   return (
     <div className="flex flex-col gap-8 py-2">
-      <TotalDeposits vault={vault} />
+      <Overview vault={vault} />
 
       <StrategyDiagram vault={vault} />
