+++ apps/evm/src/containers/VaultCard/VaultModal/PositionTab/types.ts
@@ -1,0 +1,1 @@
+export type ActionMode = 'deposit' | 'withdraw' | 'redeemAtMaturity';
