+++ apps/evm/src/containers/VaultCard/utils/index.ts
@@ -1,0 +1,2 @@
+export * from './isLegacyVenusVault';
+export * from './isPendleVault';
