+++ apps/evm/src/containers/VaultCard/utils/isLegacyVenusVault/index.ts
