+++ apps/evm/src/libs/translations/translations/ja.json
@@ -1856,7 +1856,6 @@
       "textualDate": "{{date, textual}}",
       "textualWithTime": "{{date, textualWithTime}}",
       "totalDeposited": "総預金額",
-      "totalStaked": "総ステーキング量",
       "youDeposited": "預入済み",
       "youStaked": "ステーキング済み"
     },
@@ -1871,6 +1870,7 @@
       "inputPlaceholder": "資産を検索",
       "others": "その他",
       "paused": "停止中",
+      "pending": "保留中",
       "refund": "返金",
       "repaying": "返済中",
       "stablecoins": "ステーブルコイン",
@@ -1889,7 +1889,7 @@
       "deposit": "ステーキング",
       "depositButtonDisabledLabel": "入金する有効な金額を入力",
       "effectiveFixedApr": "実質固定APR",
-      "effectiveFixedAprPendleTooltip": "このボールトストラテジーが提供する年率換算の固定利回りで、複利を含む実質年利率（APR）として表されます。",
+      "effectiveFixedAprTooltip": "このボールト戦略が提供する年率換算の固定利回りで、複利を含む実質年率（APR）として表されます。",
       "error": {
         "amountTooLow": "入力金額が少なすぎます。",
         "higherThanAvailable": "入力金額があなたの利用可能な{{tokenSymbol}}残高を超えています。金額を減らすか、「最大」をクリックして全残高を使用してください。",
@@ -1915,6 +1915,7 @@
         "riskDisclosureText": "Pendle は、プリンシパルトークン（PT）を通じて固定利回りを実現する分散型イールドトレーディングプロトコルです。Venus 固定金利ボールトは、Pendle のインフラストラクチャを通じて PT を取得し、Venus Core に担保として供給します。<br />Pendle があなたの預金にアクセスしたり移動させたりすることはできませんが、このボールトを利用することで両方のプロトコルからのリスクにさらされます。これにはスマートコントラクトの脆弱性、オラクルの障害、いずれかのプロトコルが悪用または停止される可能性が含まれます。満期前の引き出しは市場価格で実行され、預入額を下回る場合があります。固定 APY は預入時にロックされ、プロトコルレベルの障害に対する保護にはなりません。",
         "strategy": {
           "pendleRouter": "Pendle Router",
+          "supply": "供給",
           "users": "ユーザー",
           "venusCore": "Venus Core"
         },
