+++ apps/evm/src/libs/translations/translations/th.json
@@ -1856,7 +1856,6 @@
       "textualDate": "{{date, textual}}",
       "textualWithTime": "{{date, textualWithTime}}",
       "totalDeposited": "ยอดฝากทั้งหมด",
-      "totalStaked": "ยอดสเตคทั้งหมด",
       "youDeposited": "คุณฝากแล้ว",
       "youStaked": "คุณสเตคแล้ว"
     },
@@ -1871,6 +1870,7 @@
       "inputPlaceholder": "ค้นหาสินทรัพย์",
       "others": "อื่นๆ",
       "paused": "หยุดชั่วคราว",
+      "pending": "รอดำเนินการ",
       "refund": "คืนเงิน",
       "repaying": "กำลังชำระคืน",
       "stablecoins": "สเตเบิลคอยน์",
@@ -1889,7 +1889,7 @@
       "deposit": "สเตค",
       "depositButtonDisabledLabel": "ป้อนจำนวนที่ถูกต้องเพื่อฝาก",
       "effectiveFixedApr": "APR คงที่ที่มีผล",
-      "effectiveFixedAprPendleTooltip": "ผลตอบแทนคงที่รายปีที่เสนอโดยกลยุทธ์ vault นี้ แสดงเป็นอัตราร้อยละต่อปีที่มีผลซึ่งรวมดอกเบี้ยทบต้นแล้ว",
+      "effectiveFixedAprTooltip": "ผลตอบแทนคงที่ต่อปีที่เสนอโดยกลยุทธ์ vault นี้ แสดงเป็นอัตราผลตอบแทนต่อปี (APR) ที่มีผลจริงรวมดอกเบี้ยทบต้น",
       "error": {
         "amountTooLow": "จำนวนที่ป้อนต่ำเกินไป",
         "higherThanAvailable": "จำนวนเกินกว่ายอด {{tokenSymbol}} ที่ใช้ได้ของคุณ ลดจำนวนหรือคลิก 'สูงสุด' เพื่อใช้ยอดทั้งหมด",
@@ -1915,6 +1915,7 @@
         "riskDisclosureText": "Pendle เป็นโปรโตคอลการซื้อขายผลตอบแทนแบบกระจายศูนย์ที่ช่วยให้ได้รับผลตอบแทนอัตราคงที่ผ่าน Principal Tokens (PTs) Venus Fixed Rate Vault จะส่งเงินฝากผ่านโครงสร้างพื้นฐานของ Pendle เพื่อซื้อ PT ซึ่งจะถูกนำไปเป็นหลักประกันใน Venus Core<br />แม้ว่า Pendle จะไม่สามารถเข้าถึงหรือย้ายเงินฝากของคุณได้ แต่การใช้ vault นี้ทำให้คุณเผชิญกับความเสี่ยงจากทั้งสองโปรโตคอล ซึ่งรวมถึงช่องโหว่ของสัญญาอัจฉริยะ ความล้มเหลวของ oracle และความเป็นไปได้ที่โปรโตคอลใดโปรโตคอลหนึ่งจะถูกโจมตีหรือหยุดชั่วคราว การถอนก่อนกำหนดจะดำเนินการตามราคาตลาดและอาจได้รับคืนน้อยกว่าจำนวนที่ฝาก APY คงที่จะถูกล็อคไว้เมื่อฝากและไม่สามารถป้องกันความล้มเหลวในระดับโปรโตคอล",
         "strategy": {
           "pendleRouter": "Pendle Router",
+          "supply": "ฝาก",
           "users": "ผู้ใช้งาน",
           "venusCore": "Venus Core"
         },
