+++ apps/evm/src/utilities/index.ts
@@ -60,5 +60,3 @@ export * from './calculateUserPoolValues';
 export * from './addUserBorrowLimitShares';
 export * from './clampToZero';
 export * from './invalidatePendleVaultCaches';
-export * from './isPendleVault';
-export * from './isLegacyVenusVault';
